Add GbFeedbackPanel component for the course grade override, and refactor everything that currently uses a standard FeedbackPanel, to use this one. #1369
