+++ gradebookng/tool/src/java/org/sakaiproject/gradebookng/GradebookNgApplication.properties
@@ -196,6 +196,9 @@ assignment.option.setungraded = Set Score for Empty Cells
 message.edititem.success=Gradebook item ''{0}'' has been updated.
 message.edititem.error=An error occurred updating the gradebook item.
 message.editcomment.error=An error occurred editing the comment.
+message.addcoursegradeoverride.success = The course grade was updated.
+message.addcoursegradeoverride.error = An error occurred updating the course grade.
+
 
 message.updateungradeditems.success=The ungraded items were updated.
