+++ gradebookng/tool/src/java/org/sakaiproject/gradebookng/tool/panels/CourseGradeOverridePanel.html
@@ -7,8 +7,10 @@
 		<h3 wicket:id="heading">Course Grade Override for Tony Stark (tstark1)</h3>
 		
 		<div class="instruction"><wicket:message key="coursegrade.override.instructions.1" /></div>
-		
+				
 		<form wicket:id="form">
+		
+			<div wicket:id="feedback"></div>
 		
 			<table class="table table-bordered table-striped table-hover table-condensed">
 				<thead>
